implement max starboards

+++ src/constants.rs
@@ -37,7 +37,7 @@ pub const MAX_NEWER_THAN: i64 = 86_400 * 7; // one week
 pub const MAX_OLDER_THAN: i64 = MAX_NEWER_THAN;
 
 pub const MAX_VOTE_EMOJIS: usize = 3;
-pub const MAX_STARBOARDS: i16 = 3; // todo
+pub const MAX_STARBOARDS: i64 = 3;
 
 // Override Validation
 pub const MAX_CHANNELS_PER_OVERRIDE: usize = 2;
